Initial Octo SPI support
Much of the functionality shared with QuadSPI
